[tests] Increased the amount of data for tst_QSparqlTrackerDirect::destroy_connection_partially_iterated_results. Fixes MER#1192
This is a flaky test which depends on the speed of
the HW in which it is running.

In fast HW, too few data in the test won't give us
the chance to destroy the connection before the
query has finished.

In slow HW, too much data may timeout before we
start to have the results of the query.