[libqtsparql] Avoid QDBusInterface on tracker driver. Contributes to JB#49711
QDBusInterface was causing problems with the more recent
Tracker versions which do autoquit on idle since 2.2.

First, if QSparqlConnection was causing auto-start of the service,
QDBusInterface::isValid() was incorrectly reporting invalid interface
and applications using this broke.

Quoting the Qt docs: "Note: when dealing with remote objects, it is not
always possible to determine if it exists when creating a
QDBusInterface."

Second, even without the isValid check instantiating QDBusInterface
blocks and can cause futher breakage. If tracker-store is not starting
at all and app using this was autostarted with d-bus too, the result
was that the application autostarting d-bus command timed out.

So let's just use manually constructed d-bus messages and interpret
connection validity as the local side of it. Executing queries will
anyway report failures from the remote side and that will happen
asynchronously.