[libqtsparql] Remove error check that return false positives. Contributes to JB#49711
Not entirely unproblematic, but isValid() was preventing the tracker
driver to be used with d-bus autostarting tracker-store service.

Quoting the Qt docs: "Note: when dealing with remote objects, it is not
always possible to determine if it exists when creating a
QDBusInterface."

Maybe just best to return valid state here and let query execution
fail. The service might anyway die out in the middle of a session and
never come back, and such doesn't change the validity of this driver.